projects/capstone/cap_project/requirements.txt
alembic==1.4.3
aniso8601==6.0.0
astroid==2.2.5
certifi==2020.6.20
Click==7.0
ecdsa==0.13.3
Flask==1.0.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
future==0.17.1
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.0
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.2
pycryptodome==3.6.6
pylint==2.3.1
python-dateutil==2.8.1
python-editor==1.0.4
python-jose-cryptodome==1.3.2
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.3
typed-ast==1.3.5
Werkzeug==0.15.3
wrapt==1.11.1
